Added generator strings and options, misc other things
* added a generator strings feature to toggle amplifiedness
* added a GUI system for singleplayer world customization
* fixed typos, improved debug output
* improved code comments
* also made code comments worse

The current generator strings setup is uncomplete. I plan to
add various more options to customize the world, the two current
ones are just the beginning. Setting them to -1 makes everything
same as normal, -2 is a preset for 'amplified' values, and
the user can input custom values if they'd like.

Generator options are seperated by the # symbol, quasi-following
the precedence set up by ted80. Currently, the values are bytes,
but floats and strings will likely be supported values in the
future.

Yes code is ugly, will clean up unused code and comments in
subsequent commits.

Signed-off-by: Richard Rogalski <rrogalski@firemail.cc>